[form-factors] Move DM2016 parametrization to separate files
