+++ eos/form-factors/Makefile.am
@@ -38,6 +38,7 @@ libeosformfactors_la_SOURCES = \
 	parametric-bmrvd2022.hh parametric-bmrvd2022-impl.hh parametric-bmrvd2022.cc \
 	parametric-bsz2015.hh parametric-bsz2015-impl.hh parametric-bsz2015.cc \
 	parametric-dkmr2017.hh parametric-dkmr2017-impl.hh parametric-dkmr2017.cc \
+	parametric-dm2016.hh parametric-dm2016-impl.hh parametric-dm2016.cc \
 	parametric-fvdv2018.hh parametric-fvdv2018-impl.hh parametric-fvdv2018.cc \
 	parametric-kkvdz2022.hh parametric-kkvdz2022.cc \
 	parametric-kmpw2010.hh parametric-kmpw2010-impl.hh parametric-kmpw2010.cc \
@@ -71,6 +72,7 @@ include_eos_rarebdecays_HEADERS = \
 	parametric-bgjvd2019.hh \
 	parametric-bsz2015.hh \
 	parametric-dkmr2017.hh \
+	parametric-dm2016.hh \
 	parametric-fvdv2018.hh \
 	parametric-kmpw2010.hh \
 	pi-lcdas.hh
@@ -100,6 +102,7 @@ TESTS = \
 	parametric-bmrvd2022_TEST \
 	parametric-bsz2015_TEST \
 	parametric-dkmr2017_TEST \
+	parametric-dm2016_TEST \
 	parametric-fvdv2018_TEST \
 	parametric-kkvdz2022_TEST \
 	parametric-kmpw2010_TEST \
@@ -157,6 +160,8 @@ parametric_bsz2015_TEST_SOURCES = parametric-bsz2015_TEST.cc
 parametric_bmrvd2022_TEST_SOURCES = parametric-bmrvd2022_TEST.cc
 
 parametric_dkmr2017_TEST_SOURCES = parametric-dkmr2017_TEST.cc
+
+parametric_dm2016_TEST_SOURCES = parametric-dm2016_TEST.cc
 
 parametric_fvdv2018_TEST_SOURCES = parametric-fvdv2018_TEST.cc
