+++ eos/form-factors/baryonic-impl.hh
@@ -39,177 +39,6 @@ namespace eos
      * J=1/2^+ -> J=1/2^+ transitions
      */
 
-    template <typename Process_> class DM2016FormFactors :
-        public FormFactors<OneHalfPlusToOneHalfPlus>
-    {
-        private:
-            UsedParameter _alpha_0_time_v, _alpha_1_time_v, _alpha_2_time_v;
-            UsedParameter _alpha_0_time_a, _alpha_1_time_a, _alpha_2_time_a;
-
-            UsedParameter _alpha_0_long_v, _alpha_1_long_v, _alpha_2_long_v;
-            UsedParameter _alpha_0_long_a, _alpha_1_long_a, _alpha_2_long_a;
-            UsedParameter _alpha_0_perp_v, _alpha_1_perp_v, _alpha_2_perp_v;
-            UsedParameter                  _alpha_1_perp_a, _alpha_2_perp_a;
-
-            UsedParameter _alpha_0_long_t,  _alpha_1_long_t,  _alpha_2_long_t;
-            UsedParameter _alpha_0_long_t5, _alpha_1_long_t5, _alpha_2_long_t5;
-            UsedParameter _alpha_0_perp_t,  _alpha_1_perp_t,  _alpha_2_perp_t;
-            UsedParameter                   _alpha_1_perp_t5, _alpha_2_perp_t5;
-
-            static constexpr double _z(const double & t, const double & tp, const double & t0)
-            {
-                return (std::sqrt(tp - t) - std::sqrt(tp - t0)) / (std::sqrt(tp - t) + std::sqrt(tp - t0));
-            }
-
-        public:
-            DM2016FormFactors(const Parameters & p, const Options &) :
-                // time, V
-                _alpha_0_time_v(p[stringify(Process_::label) + "::a_0_time^V@DM2016"], *this),
-                _alpha_1_time_v(p[stringify(Process_::label) + "::a_1_time^V@DM2016"], *this),
-                _alpha_2_time_v(p[stringify(Process_::label) + "::a_2_time^V@DM2016"], *this),
-                // time, A
-                _alpha_0_time_a(p[stringify(Process_::label) + "::a_0_time^A@DM2016"], *this),
-                _alpha_1_time_a(p[stringify(Process_::label) + "::a_1_time^A@DM2016"], *this),
-                _alpha_2_time_a(p[stringify(Process_::label) + "::a_2_time^A@DM2016"], *this),
-
-                // long, V
-                _alpha_0_long_v(p[stringify(Process_::label) + "::a_0_long^V@DM2016"], *this),
-                _alpha_1_long_v(p[stringify(Process_::label) + "::a_1_long^V@DM2016"], *this),
-                _alpha_2_long_v(p[stringify(Process_::label) + "::a_2_long^V@DM2016"], *this),
-                // long, A
-                _alpha_0_long_a(p[stringify(Process_::label) + "::a_0_long^A@DM2016"], *this),
-                _alpha_1_long_a(p[stringify(Process_::label) + "::a_1_long^A@DM2016"], *this),
-                _alpha_2_long_a(p[stringify(Process_::label) + "::a_2_long^A@DM2016"], *this),
-                // perp, V
-                _alpha_0_perp_v(p[stringify(Process_::label) + "::a_0_perp^V@DM2016"], *this),
-                _alpha_1_perp_v(p[stringify(Process_::label) + "::a_1_perp^V@DM2016"], *this),
-                _alpha_2_perp_v(p[stringify(Process_::label) + "::a_2_perp^V@DM2016"], *this),
-                // perp, A
-                _alpha_1_perp_a(p[stringify(Process_::label) + "::a_1_perp^A@DM2016"], *this),
-                _alpha_2_perp_a(p[stringify(Process_::label) + "::a_2_perp^A@DM2016"], *this),
-
-                // long, T
-                _alpha_0_long_t(p[stringify(Process_::label) + "::a_0_long^T@DM2016"], *this),
-                _alpha_1_long_t(p[stringify(Process_::label) + "::a_1_long^T@DM2016"], *this),
-                _alpha_2_long_t(p[stringify(Process_::label) + "::a_2_long^T@DM2016"], *this),
-                // long, T5
-                _alpha_0_long_t5(p[stringify(Process_::label) + "::a_0_long^T5@DM2016"], *this),
-                _alpha_1_long_t5(p[stringify(Process_::label) + "::a_1_long^T5@DM2016"], *this),
-                _alpha_2_long_t5(p[stringify(Process_::label) + "::a_2_long^T5@DM2016"], *this),
-                // perp, T
-                _alpha_0_perp_t(p[stringify(Process_::label) + "::a_0_perp^T@DM2016"], *this),
-                _alpha_1_perp_t(p[stringify(Process_::label) + "::a_1_perp^T@DM2016"], *this),
-                _alpha_2_perp_t(p[stringify(Process_::label) + "::a_2_perp^T@DM2016"], *this),
-                // perp, T5
-                _alpha_1_perp_t5(p[stringify(Process_::label) + "::a_1_perp^T5@DM2016"], *this),
-                _alpha_2_perp_t5(p[stringify(Process_::label) + "::a_2_perp^T5@DM2016"], *this)
-            {
-            }
-
-            static FormFactors<OneHalfPlusToOneHalfPlus> * make(const Parameters & parameters, const Options & options)
-            {
-                return new DM2016FormFactors(parameters, options);
-            }
-
-            // vector current
-            virtual double f_time_v(const double & s) const
-            {
-                static const double mR2 = Process_::mR2_0p;
-
-                const double z = _z(s, Process_::tp, Process_::tm), z2 = z * z;
-
-                return 1.0 / (1.0 - s / mR2) * (_alpha_0_time_v() + _alpha_1_time_v() * z + _alpha_2_time_v() * z2);
-            }
-
-            virtual double f_long_v(const double & s) const
-            {
-                static const double mR2 = Process_::mR2_1m;
-
-                const double z = _z(s, Process_::tp, Process_::tm), z2 = z * z;
-
-                return 1.0 / (1.0 - s / mR2) * (_alpha_0_long_v() + _alpha_1_long_v() * z + _alpha_2_long_v() * z2);
-            }
-
-            virtual double f_perp_v(const double & s) const
-            {
-                static const double mR2 = Process_::mR2_1m;
-
-                const double z = _z(s, Process_::tp, Process_::tm), z2 = z * z;
-
-                return 1.0 / (1.0 - s / mR2) * (_alpha_0_perp_v() + _alpha_1_perp_v() * z + _alpha_2_perp_v() * z2);
-            }
-
-            // axial vector current
-            virtual double f_time_a(const double & s) const
-            {
-                static const double mR2 = Process_::mR2_0m;
-
-                const double z = _z(s, Process_::tp, Process_::tm), z2 = z * z;
-
-                return 1.0 / (1.0 - s / mR2) * (_alpha_0_time_a() + _alpha_1_time_a() * z + _alpha_2_time_a() * z2);
-            }
-
-            virtual double f_long_a(const double & s) const
-            {
-                static const double mR2 = Process_::mR2_1p;
-
-                const double z = _z(s, Process_::tp, Process_::tm), z2 = z * z;
-
-                return 1.0 / (1.0 - s / mR2) * (_alpha_0_long_a() + _alpha_1_long_a() * z + _alpha_2_long_a() * z2);
-            }
-
-            virtual double f_perp_a(const double & s) const
-            {
-                static const double mR2 = Process_::mR2_1p;
-
-                const double z = _z(s, Process_::tp, Process_::tm), z2 = z * z;
-
-                // Using alpha_0_long_a instead of alpha_0_perp_a, in order to
-                // fulfill relation eq. (7), [DM2016], p. 3.
-                return 1.0 / (1.0 - s / mR2) * (_alpha_0_long_a() + _alpha_1_perp_a() * z + _alpha_2_perp_a() * z2);
-            }
-
-            // tensor current
-            virtual double f_long_t(const double & s) const
-            {
-                static const double mR2 = Process_::mR2_1m;
-
-                const double z = _z(s, Process_::tp, Process_::tm), z2 = z * z;
-
-                return 1.0 / (1.0 - s / mR2) * (_alpha_0_long_t() + _alpha_1_long_t() * z + _alpha_2_long_t() * z2);
-            }
-
-            virtual double f_perp_t(const double & s) const
-            {
-                static const double mR2 = Process_::mR2_1m;
-
-                const double z = _z(s, Process_::tp, Process_::tm), z2 = z * z;
-
-                return 1.0 / (1.0 - s / mR2) * (_alpha_0_perp_t() + _alpha_1_perp_t() * z + _alpha_2_perp_t() * z2);
-            }
-
-            // axial tensor current
-            virtual double f_long_t5(const double & s) const
-            {
-                static const double mR2 = Process_::mR2_1p;
-
-                const double z = _z(s, Process_::tp, Process_::tm), z2 = z * z;
-
-                return 1.0 / (1.0 - s / mR2) * (_alpha_0_long_t5() + _alpha_1_long_t5() * z + _alpha_2_long_t5() * z2);
-            }
-
-            virtual double f_perp_t5(const double & s) const
-            {
-                static const double mR2 = Process_::mR2_1p;
-
-                const double z = _z(s, Process_::tp, Process_::tm), z2 = z * z;
-
-                // Using alpha_0_long_t5 instead of alpha_0_perp_t5, in order to
-                // fulfill relation eq. (8), [DM2016], p. 3.
-                return 1.0 / (1.0 - s / mR2) * (_alpha_0_long_t5() + _alpha_1_perp_t5() * z + _alpha_2_perp_t5() * z2);
-            }
-    };
-
     template <typename Transition_, typename Process_> class HQETFormFactors;
 
     /*
